Rename qrc20_script.rs to script_pubkey.rs
